Refactor Nix code to use an explicit pin of nixpkgs-17.09
The motivation behind introducing a `nixpkgs-17.09` pin is to
stabilize the transitive closure this codebase is expected to build
successfully with. Using the ambient `nixpkgs` is error-prone because
not all users have the exact version of `nixpkgs` this project is
built with available in their `NIX_PATH`.

Additionally, the `nixfromnpm` derivation has been restructured so
that it is provided as a package of an overridden nixpkgs instantiated
with `allowUnfree = true`, which fixes an installation error related
to a dependency that isn't completely free.

These refactors also move the `nixfromnpm` derivation modifications to
an override in the `release.nix` so that it is easier to mechanically
re-generate the `default.nix` using `cabal2nix`.