Define nodePackages as an extensible fixed point
We also make the `generatePackages` utility function overridable with
`makeOverridable` so that users can supply their own overrides
following a similar convention to that of other language package sets.

This change _should_ be backwards compatible and we also honor the
`extensions` argument with the caveat that anything provided via
`overrides` will override packages given by the `extensions` argument
to `generatePackages`.