Fix broken newScope which provided the wrong `nodePackages`
In 76e82f1 we produced the wrong `callPackage` with `newScope` because
the `nodePackages` and `self` (which is also a `nodePackages` set)
attribute sets were provided at the top-level instead of within a
`nodePackages` attribute.

This caused the `nodePackages` attribute that `callPackage` would
fill-in when calling any package to _not_ have the default `nixpkgs`
provided `nodePackages` instead of the `nodePackages` fix point we are
inside of.

This change fixes that problem.