Unset NIX_LDFLAGS, fixing "Argument list too long" errors
I suspect this doesn't correctly fix #132 but I did want to open a PR
with our (Awake Security) temporary workaround so that anyone else
with this problem can find the temporary solution and so that we can
discuss whether this suffices or if we want to approach solving this
with the suggested `passAsFile` approach.

Credit for this workaround goes entirely to @intractable (Joel
Stanley).