buildNodePackage: always make the module path directory before links
This change fixes a problem that can occur when creating the symlinks
for the circular dependencies of a package that uses a namespaced
package path.

Since we always called `mkdir -p ${dep.modulePath}` for required
dependencies before creating symlinks, we should do the same when
creating symlinks for circular dependencies.